feat: add bounce animation on project added to ballot in flow councils
