+++ src/app/flow-councils/[chainId]/[councilId]/flow-council.tsx
@@ -4,10 +4,8 @@ import { useState, useRef, useMemo, useEffect, useCallback } from "react";
 import Link from "next/link";
 import { Address } from "viem";
 import { useAccount, useSwitchChain } from "wagmi";
-import { PieChart, Pie, Cell, ResponsiveContainer } from "recharts";
 import Container from "react-bootstrap/Container";
 import Stack from "react-bootstrap/Stack";
-import Button from "react-bootstrap/Button";
 import Modal from "react-bootstrap/Modal";
 import Spinner from "react-bootstrap/Spinner";
 import Dropdown from "react-bootstrap/Dropdown";
@@ -17,11 +15,12 @@ import RoundBanner from "../../components/RoundBanner";
 import GranteeDetails from "../../components/GranteeDetails";
 import Ballot from "../../components/Ballot";
 import DistributionPoolFunding from "../../components/DistributionPoolFunding";
-import InfoTooltip from "@/components/InfoTooltip";
+import VoteBubble from "../../components/VoteBubble";
 import { ProjectMetadata } from "@/types/project";
 import { Grantee, SortingMethod } from "../../types/grantee";
 import { useMediaQuery } from "@/hooks/mediaQuery";
 import useCouncil from "../../hooks/council";
+import useAnimateVoteBubble from "../../hooks/animateVoteBubble";
 import { networks } from "@/lib/networks";
 import { shuffle, getPlaceholderImageSrc, generateColor } from "@/lib/utils";
 
@@ -40,8 +39,6 @@ export default function FlowCouncil({
   const [showDistributionPoolFunding, setShowDistributionPoolFunding] =
     useState(false);
   const [showConnectionModal, setShowConnectionModal] = useState(false);
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const [sharedPieData, setSharedPieData] = useState<any[]>([]);
 
   const skipGrantees = useRef(0);
   const hasNextGrantee = useRef(true);
@@ -63,6 +60,7 @@ export default function FlowCouncil({
     token,
     dispatchNewAllocation,
   } = useCouncil();
+  const { voteBubbleRef, animateVoteBubble } = useAnimateVoteBubble();
 
   const poolMember = gdaPool?.poolMembers.find(
     (member: { account: { id: string } }) =>
@@ -74,13 +72,6 @@ export default function FlowCouncil({
       ? currentAllocation.votingPower
       : 0;
   const currentAllocationStringified = JSON.stringify(currentAllocation);
-  const totalAllocatedProjects =
-    newAllocation?.allocation.filter((allocation) => allocation.amount !== 0)
-      .length ?? 0;
-  const totalAllocatedVotes =
-    newAllocation?.allocation
-      ?.map((allocation) => allocation.amount)
-      .reduce((a, b) => a + b, 0) ?? 0;
 
   const getGrantee = useCallback(
     (recipient: { id: string; address: string; metadata: ProjectMetadata }) => {
@@ -245,58 +236,6 @@ export default function FlowCouncil({
     setGrantees((prev) => sortGrantees(prev));
   }, [sortingMethod, sortGrantees]);
 
-  const createConsistentPieData = useCallback(() => {
-    if (!grantees.length) {
-      return [];
-    }
-
-    const allocatedVotes = newAllocation?.allocation
-      ? newAllocation.allocation.reduce((sum, a) => sum + a.amount, 0)
-      : 0;
-    const unallocatedVotes = votingPower - allocatedVotes;
-    const data = grantees.map((grantee) => {
-      const allocation = newAllocation?.allocation?.find(
-        (a) => a.grantee === grantee.address,
-      );
-
-      return {
-        id: grantee.address,
-        name: grantee.metadata.title,
-        value: allocation ? allocation.amount : 0,
-        color:
-          allocation && allocation.amount > 0
-            ? granteeColors[grantee.address]
-            : "#e0e0e0",
-      };
-    });
-
-    if (newAllocation?.allocation) {
-      newAllocation.allocation.forEach((allocation) => {
-        if (data.some((item) => item.id === allocation.grantee)) {
-          return;
-        }
-
-        data.push({
-          id: allocation.grantee,
-          name: allocation.grantee.substring(0, 6),
-          value: allocation.amount,
-          color: granteeColors[allocation.grantee] || "#1f77b4",
-        });
-      });
-    }
-
-    if (unallocatedVotes > 0) {
-      data.push({
-        id: "0xdead",
-        name: "Unallocated",
-        value: unallocatedVotes,
-        color: "#e0e0e0",
-      });
-    }
-
-    return data.filter((entry) => entry.value > 0 || entry.id === "0xdead");
-  }, [grantees, newAllocation?.allocation, granteeColors, votingPower]);
-
   const clearUnallocated = () => {
     const zeroAllocations = newAllocation?.allocation.filter(
       (a) => a.amount === 0,
@@ -312,66 +251,7 @@ export default function FlowCouncil({
     }
   };
 
-  useEffect(() => {
-    setSharedPieData(createConsistentPieData());
-  }, [createConsistentPieData]);
-
   useEffect(() => setShowConnectionModal(shouldConnect), [shouldConnect]);
-
-  const VoteButton = () => {
-    return (
-      <Button
-        className="btn btn-primary d-flex align-items-center gap-2 py-3 px-4 shadow-lg rounded-pill"
-        onClick={() => dispatchNewAllocation({ type: "show-ballot" })}
-        style={{
-          width: isMobile ? 360 : 400,
-          transition: "all 0.2s ease-in-out",
-          transform: "scale(1)",
-          boxShadow: "0 8px 16px rgba(0, 0, 0, 0.2)",
-        }}
-        onMouseEnter={(e) => {
-          e.currentTarget.style.transform = "scale(1.05)";
-          e.currentTarget.style.boxShadow = "0 12px 20px rgba(0, 0, 0, 0.25)";
-        }}
-        onMouseLeave={(e) => {
-          e.currentTarget.style.transform = "scale(1)";
-          e.currentTarget.style.boxShadow = "0 8px 16px rgba(0, 0, 0, 0.2)";
-        }}
-      >
-        <Stack
-          direction="horizontal"
-          className="justify-content-center align-items-center bg-white rounded-circle"
-          style={{ width: 64, height: 64, overflow: "hidden" }}
-        >
-          <ResponsiveContainer width="100%" height="100%">
-            <PieChart>
-              {sharedPieData ? (
-                <Pie data={sharedPieData} dataKey="value" outerRadius={50}>
-                  {sharedPieData.map((entry, index) => (
-                    <Cell key={`cell-${index}`} fill={entry.color} />
-                  ))}
-                </Pie>
-              ) : null}
-            </PieChart>
-          </ResponsiveContainer>
-        </Stack>
-        <Stack direction="vertical">
-          <span className="fs-4 fw-semibold d-block text-center">VOTE</span>
-          <small
-            className={`fs-6 d-block ${totalAllocatedVotes > votingPower ? "text-warning" : "text-white-50"}`}
-          >
-            {totalAllocatedProjects}
-            {council?.maxAllocationsPerMember
-              ? `/${council.maxAllocationsPerMember}`
-              : ""}{" "}
-            Project{totalAllocatedProjects > 1 ? "s" : ""},{" "}
-            {totalAllocatedVotes}/{votingPower} Vote
-            {totalAllocatedVotes > 1 ? "s" : ""}
-          </small>
-        </Stack>
-      </Button>
-    );
-  };
 
   return (
     <>
@@ -466,6 +346,7 @@ export default function FlowCouncil({
                 votingPower={votingPower}
                 showGranteeDetails={() => setShowGranteeDetails(grantee)}
                 granteeColor={granteeColors[grantee.address]}
+                onAddToBallot={animateVoteBubble}
               />
             ))}
           </div>
@@ -523,27 +404,12 @@ export default function FlowCouncil({
         </Modal.Footer>
       </Modal>
       {newAllocation?.allocation && newAllocation.allocation.length > 0 && (
-        <Stack
-          direction="horizontal"
-          className="position-fixed"
-          style={{
-            width: isMobile ? "100%" : "auto",
-            justifyContent: isMobile ? "center" : "right",
-            bottom: "2rem",
-            right: isMobile ? "auto" : "2rem",
-            zIndex: 3,
-          }}
-        >
-          {isMobile || isTablet ? (
-            <VoteButton />
-          ) : (
-            <InfoTooltip
-              position={{ top: true }}
-              content={<>Click to edit & submit your votes</>}
-              target={<VoteButton />}
-            />
-          )}
-        </Stack>
+        <VoteBubble
+          grantees={grantees}
+          granteeColors={granteeColors}
+          votingPower={votingPower}
+          voteBubbleRef={voteBubbleRef}
+        />
       )}
     </>
   );
