+++ src/app/gooddollar/good-dollar.tsx
@@ -8,19 +8,18 @@ import Container from "react-bootstrap/Container";
 import Stack from "react-bootstrap/Stack";
 import Modal from "react-bootstrap/Modal";
 import Spinner from "react-bootstrap/Spinner";
-import Button from "react-bootstrap/Button";
 import Dropdown from "react-bootstrap/Dropdown";
-import { PieChart, Pie, Cell, ResponsiveContainer } from "recharts";
 import PoolConnectionButton from "@/components/PoolConnectionButton";
 import GranteeCard from "@/app/flow-councils/components/GranteeCard";
 import RoundBanner from "./components/RoundBanner";
-import InfoTooltip from "@/components/InfoTooltip";
 import Ballot from "./components/Ballot";
 import GranteeDetails from "@/app/flow-councils/components/GranteeDetails";
+import VoteBubble from "@/app/flow-councils/components/VoteBubble";
 import DistributionPoolFunding from "./components/DistributionPoolFunding";
 import { ProjectMetadata } from "@/types/project";
 import { Grantee, SortingMethod } from "@/app/flow-councils/types/grantee";
 import useCouncil from "@/app/flow-councils/hooks/council";
+import useAnimateVoteBubble from "@/app/flow-councils/hooks/animateVoteBubble";
 import { useMediaQuery } from "@/hooks/mediaQuery";
 import { networks } from "@/lib/networks";
 import { shuffle, getPlaceholderImageSrc, generateColor } from "@/lib/utils";
@@ -35,8 +34,6 @@ export default function GoodDollar({ chainId }: { chainId: number }) {
   const [showDistributionPoolFunding, setShowDistributionPoolFunding] =
     useState(false);
   const [showConnectionModal, setShowConnectionModal] = useState(false);
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const [sharedPieData, setSharedPieData] = useState<any[]>([]);
 
   const skipGrantees = useRef(0);
   const hasNextGrantee = useRef(true);
@@ -58,6 +55,7 @@ export default function GoodDollar({ chainId }: { chainId: number }) {
     token,
     dispatchNewAllocation,
   } = useCouncil();
+  const { voteBubbleRef, animateVoteBubble } = useAnimateVoteBubble();
 
   const poolMember = gdaPool?.poolMembers.find(
     (member: { account: { id: string } }) =>
@@ -69,13 +67,6 @@ export default function GoodDollar({ chainId }: { chainId: number }) {
       ? currentAllocation.votingPower
       : 0;
   const currentAllocationStringified = JSON.stringify(currentAllocation);
-  const totalAllocatedProjects =
-    newAllocation?.allocation.filter((allocation) => allocation.amount !== 0)
-      .length ?? 0;
-  const totalAllocatedVotes =
-    newAllocation?.allocation
-      ?.map((allocation) => allocation.amount)
-      .reduce((a, b) => a + b, 0) ?? 0;
 
   const getGrantee = useCallback(
     (recipient: {
@@ -244,62 +235,6 @@ export default function GoodDollar({ chainId }: { chainId: number }) {
     }
   }, [address, connectedChain, chainId, switchChain]);
 
-  const createConsistentPieData = useCallback(() => {
-    if (!grantees.length) {
-      return [];
-    }
-
-    const allocatedVotes = newAllocation?.allocation
-      ? newAllocation.allocation.reduce((sum, a) => sum + a.amount, 0)
-      : 0;
-    const unallocatedVotes = votingPower - allocatedVotes;
-    const data = grantees.map((grantee) => {
-      const allocation = newAllocation?.allocation?.find(
-        (a) => a.grantee === grantee.address,
-      );
-
-      return {
-        id: grantee.address,
-        name: grantee.metadata.title,
-        value: allocation ? allocation.amount : 0,
-        color:
-          allocation && allocation.amount > 0
-            ? granteeColors[grantee.address]
-            : "#e0e0e0",
-      };
-    });
-
-    if (newAllocation?.allocation) {
-      newAllocation.allocation.forEach((allocation) => {
-        if (data.some((item) => item.id === allocation.grantee)) {
-          return;
-        }
-
-        data.push({
-          id: allocation.grantee,
-          name: allocation.grantee.substring(0, 6),
-          value: allocation.amount,
-          color: granteeColors[allocation.grantee] || "#1f77b4",
-        });
-      });
-    }
-
-    if (unallocatedVotes > 0) {
-      data.push({
-        id: "0xdead",
-        name: "Unallocated",
-        value: unallocatedVotes,
-        color: "#e0e0e0",
-      });
-    }
-
-    return data.filter((entry) => entry.value > 0 || entry.id === "0xdead");
-  }, [grantees, newAllocation?.allocation, granteeColors, votingPower]);
-
-  useEffect(() => {
-    setSharedPieData(createConsistentPieData());
-  }, [createConsistentPieData]);
-
   useEffect(() => setShowConnectionModal(shouldConnect), [shouldConnect]);
 
   const clearUnallocated = () => {
@@ -315,61 +250,6 @@ export default function GoodDollar({ chainId }: { chainId: number }) {
         });
       }
     }
-  };
-
-  const VoteButton = () => {
-    return (
-      <Button
-        className="btn btn-primary d-flex align-items-center gap-2 py-3 px-4 shadow-lg rounded-pill"
-        onClick={() => dispatchNewAllocation({ type: "show-ballot" })}
-        style={{
-          width: isMobile ? 360 : 400,
-          transition: "all 0.2s ease-in-out",
-          transform: "scale(1)",
-          boxShadow: "0 8px 16px rgba(0, 0, 0, 0.2)",
-        }}
-        onMouseEnter={(e) => {
-          e.currentTarget.style.transform = "scale(1.05)";
-          e.currentTarget.style.boxShadow = "0 12px 20px rgba(0, 0, 0, 0.25)";
-        }}
-        onMouseLeave={(e) => {
-          e.currentTarget.style.transform = "scale(1)";
-          e.currentTarget.style.boxShadow = "0 8px 16px rgba(0, 0, 0, 0.2)";
-        }}
-      >
-        <Stack
-          direction="horizontal"
-          className="justify-content-center align-items-center bg-white rounded-circle"
-          style={{ width: 64, height: 64, overflow: "hidden" }}
-        >
-          <ResponsiveContainer width="100%" height="100%">
-            <PieChart>
-              {sharedPieData ? (
-                <Pie data={sharedPieData} dataKey="value" outerRadius={50}>
-                  {sharedPieData.map((entry, index) => (
-                    <Cell key={`cell-${index}`} fill={entry.color} />
-                  ))}
-                </Pie>
-              ) : null}
-            </PieChart>
-          </ResponsiveContainer>
-        </Stack>
-        <Stack direction="vertical">
-          <span className="fs-4 fw-semibold d-block text-center">VOTE</span>
-          <small
-            className={`fs-6 d-block ${totalAllocatedVotes > votingPower ? "text-warning" : "text-white-50"}`}
-          >
-            {totalAllocatedProjects}
-            {council?.maxAllocationsPerMember
-              ? `/${council.maxAllocationsPerMember}`
-              : ""}{" "}
-            Project{totalAllocatedProjects > 1 ? "s" : ""},{" "}
-            {totalAllocatedVotes}/{votingPower} Vote
-            {totalAllocatedVotes > 1 ? "s" : ""}
-          </small>
-        </Stack>
-      </Button>
-    );
   };
 
   return (
@@ -466,6 +346,7 @@ export default function GoodDollar({ chainId }: { chainId: number }) {
                 votingPower={votingPower}
                 showGranteeDetails={() => setShowGranteeDetails(grantee)}
                 granteeColor={granteeColors[grantee.address]}
+                onAddToBallot={animateVoteBubble}
               />
             ))}
           </div>
@@ -523,27 +404,12 @@ export default function GoodDollar({ chainId }: { chainId: number }) {
         </Modal.Footer>
       </Modal>
       {newAllocation?.allocation && newAllocation.allocation.length > 0 && (
-        <Stack
-          direction="horizontal"
-          className="position-fixed"
-          style={{
-            width: isMobile ? "100%" : "auto",
-            justifyContent: isMobile ? "center" : "right",
-            bottom: "2rem",
-            right: isMobile ? "auto" : "2rem",
-            zIndex: 3,
-          }}
-        >
-          {isMobile || isTablet ? (
-            <VoteButton />
-          ) : (
-            <InfoTooltip
-              position={{ top: true }}
-              content={<>Click to edit & submit your votes</>}
-              target={<VoteButton />}
-            />
-          )}
-        </Stack>
+        <VoteBubble
+          grantees={grantees}
+          granteeColors={granteeColors}
+          votingPower={votingPower}
+          voteBubbleRef={voteBubbleRef}
+        />
       )}
     </>
   );
